pyproject.toml
[tool.poetry]
name = "pdf-generation"
version = "0.1.0"
description = ""
authors = ["Pedro Augusto. <pedr.augustobarbosa.aparecido@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.34.0"
playwright = "^1.49.1"
pillow = "^11.0.0"
flask = "^3.1.0"
flask-cors = "^5.0.0"
loguru = "^0.7.3"
firebase-admin = "^6.6.0"
load-dotenv = "^0.1.0"
pypdf2 = "^3.0.1"
reportlab = "^4.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
